services/crewai-service/requirements.txt
crewai>=0.28.0
anthropic>=0.5.0
fastapi>=0.104.0
uvicorn>=0.23.2
pydantic>=2.4.0
structlog>=23.1.0
google-cloud-pubsub>=2.18.0
prometheus-client>=0.17.0
httpx>=0.25.0
python-dotenv>=1.0.0
tenacity>=8.2.0
qdrant-client>=1.5.0